requirements.txt
# SPDX-FileCopyrightText: The MF-Plugin Contributors
# SPDX-License-Identifier: CC0-1.0

antlr4-python3-runtime==4.9.3
antlr-denter
croniter
snakes
